+++ apps/console-lite-e2e/src/integration/portfolio-page.test.ts
@@ -66,7 +66,7 @@ describe('Portfolio page tabs', { tags: '@smoke' }, () => {
 
     it('data should be properly rendered', () => {
       cy.get('.ag-center-cols-container .ag-row').should('have.length', 5);
-      cy.get('[title="tEURO"] button').click();
+      cy.contains('.ag-center-cols-container button', 'tEURO').click();
       cy.getByTestId('dialog-title').should(
         'have.text',
         'Asset details - tEURO'
