+++ apps/console-lite/src/app/components/portfolio/accounts/accounts.tsx
@@ -41,7 +41,7 @@ const AccountsManager = () => {
     const rowsThisBlock = dataRef.current
       ? dataRef.current.slice(startRow, endRow)
       : [];
-    const lastRow = dataRef.current?.length ?? -1;
+    const lastRow = dataRef.current ? dataRef.current.length : 0;
     successCallback(rowsThisBlock, lastRow);
   };
   const { columnDefs, defaultColDef } = useAccountColumnDefinitions();
