+++ apps/trading-e2e/src/integration/trading-portfolio.cy.ts
@@ -16,7 +16,11 @@ describe('Portfolio page', { tags: '@smoke' }, () => {
       cy.getByTestId('"Ledger entries"').click();
       const headers = [
         'Sender',
+        'Account type',
+        'Market',
         'Receiver',
+        'Account type',
+        'Market',
         'Transfer Type',
         'Quantity',
         'Asset',
