+++ apps/trading/client-pages/portfolio/withdrawals-container.tsx
@@ -1,6 +1,5 @@
 import { AsyncRenderer, Button } from '@vegaprotocol/ui-toolkit';
 import {
-  PendingWithdrawalsTable,
   useWithdrawals,
   useWithdrawalDialog,
   WithdrawalsTable,
@@ -9,35 +8,27 @@ import { t } from '@vegaprotocol/react-helpers';
 import { VegaWalletContainer } from '../../components/vega-wallet-container';
 
 export const WithdrawalsContainer = () => {
-  const { pending, completed, loading, error } = useWithdrawals();
+  const { data, loading, error } = useWithdrawals();
   const openWithdrawDialog = useWithdrawalDialog((state) => state.open);
 
   return (
     <VegaWalletContainer>
       <div className="h-full relative grid grid-rows-[1fr,min-content]">
-        <div className="h-full">
-          <AsyncRenderer
-            data={{ pending, completed }}
-            loading={loading}
-            error={error}
-            render={({ pending, completed }) => (
-              <>
-                {pending && pending.length > 0 && (
-                  <>
-                    <h4 className="pt-3 pb-1">{t('Pending withdrawals')}</h4>
-                    <PendingWithdrawalsTable rowData={pending} />
-                  </>
-                )}
-                {completed && completed.length > 0 && (
-                  <h4 className="pt-3 pb-1">{t('Withdrawal history')}</h4>
-                )}
-                <WithdrawalsTable
-                  data-testid="withdrawals-history"
-                  rowData={completed}
-                />
-              </>
-            )}
+        <div className="h-full relative">
+          <WithdrawalsTable
+            data-testid="withdrawals-history"
+            rowData={data}
+            noRowsOverlayComponent={() => null}
           />
+          <div className="pointer-events-none absolute inset-0">
+            <AsyncRenderer
+              data={data}
+              loading={loading}
+              error={error}
+              noDataCondition={(data) => !(data && data.length)}
+              noDataMessage={t('No withdrawals')}
+            />
+          </div>
         </div>
         <div className="w-full dark:bg-black bg-white absolute bottom-0 h-auto flex justify-end px-[11px] py-2">
           <Button
