+++ libs/withdraws/src/lib/withdrawals-table.spec.tsx
@@ -26,11 +26,12 @@ describe('renders the correct columns', () => {
     });
 
     const headers = screen.getAllByRole('columnheader');
-    expect(headers).toHaveLength(6);
+    expect(headers).toHaveLength(7);
     expect(headers.map((h) => h.textContent?.trim())).toEqual([
       'Asset',
       'Amount',
       'Recipient',
+      'Created',
       'Completed',
       'Status',
       'Transaction',
@@ -41,9 +42,10 @@ describe('renders the correct columns', () => {
       'asset-symbol',
       '1.00',
       '123456…123456',
+      getTimeFormat().format(new Date(withdrawal.createdTimestamp as string)),
       '-',
       'Pending',
-      '-',
+      'Complete withdrawal',
     ];
     cells.forEach((cell, i) => {
       expect(cell).toHaveTextContent(expectedValues[i]);
@@ -66,6 +68,7 @@ describe('renders the correct columns', () => {
       'asset-symbol',
       '1.00',
       '123456…123456',
+      getTimeFormat().format(new Date(withdrawal.createdTimestamp as string)),
       getTimeFormat().format(new Date(withdrawal.withdrawnTimestamp as string)),
       'Completed',
       '0x1234…121314',
